project structure defined
