+++ lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_native_splash/flutter_native_splash.dart';
+import 'package:tasker_mobile/src/features/auth/presentation/login_screen.dart';
 
 void main() {
   WidgetsBinding widgetsBinding = WidgetsFlutterBinding.ensureInitialized();
@@ -18,30 +19,8 @@ class _MyAppState extends State<MyApp> {
   @override
   Widget build(BuildContext context) {
     return const MaterialApp(
-      home: HomePage(),
-    );
-  }
-}
-
-class HomePage extends StatelessWidget {
-  const HomePage({super.key});
-
-  Future removeSplashScreen() async {
-    await Future.delayed(
-        const Duration(seconds: 5), () => FlutterNativeSplash.remove());
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    removeSplashScreen();
-
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text("Tasker"),
-      ),
-      body: const Center(
-        child: Text("Hello World"),
-      ),
+      title: 'Tasker',
+      home: LoginScreen(),
     );
   }
 }
